+++ cmd/src/campaigns_common.go
@@ -42,6 +42,7 @@ type campaignsApplyFlags struct {
 	parallelism      int
 	timeout          time.Duration
 	cleanArchives    bool
+	skipErrors       bool
 }
 
 func newCampaignsApplyFlags(flagSet *flag.FlagSet, cacheDir, tempDir string) *campaignsApplyFlags {
@@ -94,6 +95,10 @@ func newCampaignsApplyFlags(flagSet *flag.FlagSet, cacheDir, tempDir string) *ca
 	flagSet.BoolVar(
 		&caf.cleanArchives, "clean-archives", true,
 		"If true, deletes downloaded repository archives after executing campaign steps.",
+	)
+	flagSet.BoolVar(
+		&caf.skipErrors, "skip-errors", false,
+		"If true, errors encountered while executing steps in a repository won't stop the execution of the campaign spec but only cause that repository to be skipped.",
 	)
 
 	flagSet.BoolVar(verbose, "v", false, "print verbose output")
@@ -235,11 +240,16 @@ func campaignsExecute(ctx context.Context, out *output.Output, svc *campaigns.Se
 	}
 
 	p := newCampaignProgressPrinter(out, *verbose, opts.Parallelism)
-	specs, err := svc.ExecuteCampaignSpec(ctx, repos, executor, campaignSpec, p.PrintStatuses)
-	if err != nil {
+	specs, err := svc.ExecuteCampaignSpec(ctx, repos, executor, campaignSpec, p.PrintStatuses, flags.skipErrors)
+	if err != nil && !flags.skipErrors {
 		return "", "", err
+
 	}
 	p.Complete()
+	if err != nil && flags.skipErrors {
+		printExecutionError(out, err)
+		out.WriteLine(output.Line(output.EmojiWarning, output.StyleWarning, "Skipping errors because -skip-errors was used."))
+	}
 
 	if logFiles := executor.LogFiles(); len(logFiles) > 0 && flags.keepLogs {
 		func() {
@@ -317,40 +327,54 @@ func campaignsParseSpec(out *output.Output, svc *campaigns.Service, input io.Rea
 func printExecutionError(out *output.Output, err error) {
 	out.Write("")
 
-	writeErr := func(block *output.Block, err error) {
-		if block == nil {
-			return
-		}
+	writeErrs := func(errs []error) {
+		var block *output.Block
 
-		if taskErr, ok := err.(campaigns.TaskExecutionErr); ok {
-			block.Write(formatTaskExecutionErr(taskErr))
+		if len(errs) > 1 {
+			block = out.Block(output.Linef(output.EmojiFailure, output.StyleWarning, "%d errors:", len(errs)))
 		} else {
-			block.Write(err.Error())
+			block = out.Block(output.Line(output.EmojiFailure, output.StyleWarning, "Error:"))
+		}
+
+		for _, e := range errs {
+			if taskErr, ok := e.(campaigns.TaskExecutionErr); ok {
+				block.Write(formatTaskExecutionErr(taskErr))
+			} else {
+				block.Writef("%s%s", output.StyleBold, e.Error())
+			}
+		}
+
+		if block != nil {
+			block.Close()
 		}
 	}
 
-	var block *output.Block
-	singleErrHeader := output.Line(output.EmojiFailure, output.StyleWarning, "Error:")
+	switch err := err.(type) {
+	case parallel.Errors, *multierror.Error:
+		writeErrs(flattenErrs(err))
 
-	if parErr, ok := err.(parallel.Errors); ok {
-		if len(parErr) > 1 {
-			block = out.Block(output.Linef(output.EmojiFailure, output.StyleWarning, "%d errors:", len(parErr)))
-		} else {
-			block = out.Block(singleErrHeader)
-		}
-
-		for _, e := range parErr {
-			writeErr(block, e)
-		}
-	} else {
-		block = out.Block(singleErrHeader)
-		writeErr(block, err)
+	default:
+		writeErrs([]error{err})
 	}
 
-	if block != nil {
-		block.Close()
+}
+
+func flattenErrs(err error) (result []error) {
+	switch errs := err.(type) {
+	case parallel.Errors:
+		for _, e := range errs {
+			result = append(result, flattenErrs(e)...)
+		}
+
+	case *multierror.Error:
+		for _, e := range errs.Errors {
+			result = append(result, flattenErrs(e)...)
+		}
+	default:
+		result = append(result, errs)
 	}
-	out.Write("")
+
+	return result
 }
 
 func formatTaskExecutionErr(err campaigns.TaskExecutionErr) string {