+++ internal/campaigns/service.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/src-cli/internal/api"
 	"github.com/sourcegraph/src-cli/internal/campaigns/graphql"
@@ -215,7 +216,7 @@ func (svc *Service) SetDockerImages(ctx context.Context, spec *CampaignSpec, pro
 	return nil
 }
 
-func (svc *Service) ExecuteCampaignSpec(ctx context.Context, repos []*graphql.Repository, x Executor, spec *CampaignSpec, progress func([]*TaskStatus)) ([]*ChangesetSpec, error) {
+func (svc *Service) ExecuteCampaignSpec(ctx context.Context, repos []*graphql.Repository, x Executor, spec *CampaignSpec, progress func([]*TaskStatus), skipErrors bool) ([]*ChangesetSpec, error) {
 	statuses := make([]*TaskStatus, 0, len(repos))
 	for _, repo := range repos {
 		ts := x.AddTask(repo, spec.Steps, spec.ChangesetTemplate)
@@ -242,6 +243,8 @@ func (svc *Service) ExecuteCampaignSpec(ctx context.Context, repos []*graphql.Re
 		}()
 	}
 
+	var errs *multierror.Error
+
 	x.Start(ctx)
 	specs, err := x.Wait()
 	if progress != nil {
@@ -249,14 +252,24 @@ func (svc *Service) ExecuteCampaignSpec(ctx context.Context, repos []*graphql.Re
 		done <- struct{}{}
 	}
 	if err != nil {
-		return nil, err
+		if skipErrors {
+			errs = multierror.Append(errs, err)
+		} else {
+			return nil, err
+		}
 	}
 
 	// Add external changeset specs.
 	for _, ic := range spec.ImportChangesets {
 		repo, err := svc.resolveRepositoryName(ctx, ic.Repository)
 		if err != nil {
-			return nil, errors.Wrapf(err, "resolving repository name %q", ic.Repository)
+			wrapped := errors.Wrapf(err, "resolving repository name %q", ic.Repository)
+			if skipErrors {
+				errs = multierror.Append(errs, wrapped)
+				continue
+			} else {
+				return nil, wrapped
+			}
 		}
 
 		for _, id := range ic.ExternalIDs {
@@ -284,7 +297,7 @@ func (svc *Service) ExecuteCampaignSpec(ctx context.Context, repos []*graphql.Re
 		}
 	}
 
-	return specs, nil
+	return specs, errs.ErrorOrNil()
 }
 
 func (svc *Service) ParseCampaignSpec(in io.Reader) (*CampaignSpec, string, error) {